# app/openapi.yml
openapi: "3.0.0"
servers:
  - description:  Test
    url: /
info:
  description: This is a simple API
  version: "1.0.0"
  title: Test API
paths:
...
  /sale:
    post:
      tags:
        - Sale
      summary: Add a new Sale
      operationId: app.views.saleView.add_invoice
      description: |
        Add a new Invoice
      requestBody:
        description: Creates a new Sale 
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Sale'
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /sale/{dealer_id}:
    get:
      tags:
        - Sale
      summary: List all sales in a dealer for a given date range 
      operationId: app.views.saleView.get_sale_by_date
      description: |
        List all sales in a dealer for a given date range 
      parameters:
        - name: dealer_id
          in: path 
          description: Dealer ID
          required: true 
          schema:
            type: integer 
            example: 1
        - name: start_dt 
          in: query 
          description: start date for date filter 
          required: false 
          schema:
            type: string 
            example: "2023-06-13"
        - name: end_dt 
          in: query
          description: end date for date filter 
          required: false 
          schema:
            type: string 
            example: "2023-06-13" 

      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAllSaleResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /sale/{dealer_id}/{customer_id}:
    get:
      tags:
        - Sale
      summary: List all sales in a dealer for a customer in a given date range 
      operationId: app.views.saleView.get_invoice_by_cust_date
      description: |
        List all sales in a dealer for a customer in a given date range 
      parameters:
        - name: dealer_id
          in: path 
          description: Dealer ID
          required: true 
          schema:
            type: integer 
            example: 1
        - name: customer_id
          in: path 
          description: Customer ID
          required: true 
          schema:
            type: integer 
            example: 1
        - name: start_dt 
          in: query 
          description: start date for date filter 
          required: false 
          schema:
            type: string 
            example: "2023-06-13"
        - name: end_dt 
          in: query
          description: end date for date filter 
          required: false 
          schema:
            type: string 
            example: "2023-06-13" 
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAllSaleResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /sale/{invoice_id}:
    patch:
      tags:
        - Sale
      summary: Update/Patch Sale by invoice id 
      operationId: app.views.saleView.update_sale_record
      description: |
        Updates a single record based on invoice_id 
      parameters:
        - name: invoice_id
          in: path 
          description: Invoice ID
          required: true 
          schema:
            type: integer 
            example: 1
      requestBody:
        description: Updates a sale record
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SalePatch'
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

    delete:
      tags:
        - Sale
      summary: Deletes Sale by invoice id
      operationId: app.views.saleView.delete_sale_record
      description: |
        Deletes a single sale  record 
      parameters:
        - name: invoice_id
          in: path 
          description: Invoice ID
          required: true 
          schema:
            type: integer 
            example: 1
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /state_tax:
    post:
      tags:
        - State_Tax
      summary: Add new tax information for state
      operationId: app.views.stateTaxView.add_state_tax
      description: |
        Add new tax information for state
      requestBody:
        description: Add new tax information for state
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StateTax'
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    get:
      tags:
        - State_Tax
      summary: List tax information for all states
      operationId: app.views.stateTaxView.get_all_state_tax
      description: |
        List tax information for all states
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAllStateTaxResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /state_tax/{state_tax_id}:
    get:
      tags:
        - State_Tax
      summary: Get tax information by state_tax_id
      operationId: app.views.stateTaxView.get_state_by_id
      description: |
        Get tax information by state_tax_id
      parameters:
        - name: state_tax_id
          in: path 
          description: state_tax_id
          required: true 
          schema:
            type: integer 
            example: 1
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetStateTaxResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        
    patch:
      tags:
        - State_Tax
      summary: Update/Patch State Tax  by state_tax_id
      operationId: app.views.stateTaxView.update_state_tax
      description: |
        Update/Patch Sale line item  by state_tax_id
      parameters:
        - name: state_tax_id
          in: path 
          description: state_tax_id
          required: true 
          schema:
            type: integer 
            example: 1
      requestBody:
        description: Update/Patch State Tax  by state_tax_id
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StateTaxPatch'
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

    delete:
      tags:
        - State_Tax
      summary: Deletes state tax by state_tax_id
      operationId: app.views.stateTaxView.delete_state_tax
      description: |
        Deletes a single state tax  record 
      parameters:
        - name: state_tax_id
          in: path 
          description: state_tax_id
          required: true 
          schema:
            type: integer 
            example: 1
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /state_tax/state/{state}:
    get:
      tags:
        - State_Tax
      summary: List all State_Tax items for a given state 
      operationId: app.views.stateTaxView.get_state_tax
      description: |
        List all State_Tax items for a given state 
      parameters:
        - name: state
          in: path 
          description: state
          required: true 
          schema:
            type: string 
            example: california
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetStateTaxResponse'
        '400':
          description: Bad request 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
components:
  schemas:
...
    Sale:
      required: 
        - date_of_purchase
        - dealer_id
        - customer_id
        - employee_id 
      type: object
      properties:
        date_of_purchase:
          type: string
          example: "2023-06-13"
        customer_id:
          type: integer 
          example: 1
        dealer_id:
          type: integer 
          example: 1
        employee_id:
          type: integer
          example: 1 
    GetAllSaleResponse:
      type: object 
      properties:
        msg:
          type: array
          items:
            $ref: '#/components/schemas/SaleObject'
        status:
          type: string 
    SaleObject:
      type: object
      properties:
        date_of_purchase:
          type: string
          example: "Tue, 13 Jun 2023 00:00:00 GMT"
        customer_id:
          type: integer 
          example: 1
        dealer_id:
          type: integer 
          example: 1
        employee_id:
          type: integer
          example: 1 
        invoice_id:
          type: integer
          example: 1 
        net_price:
          type: number 
          example: 100.0
        net_tax:
          type: number 
          example: 30.0
        created_at:
          type: string 
          example: Mon, 12 Jun 2022 08:36:01 GMT  
        modified_at:
          type: string 
          example: Mon, 12 Jun 2022 08:36:01 GMT
...
    SalePatch:
      type: object
      properties:
        date_of_purchase:
          type: string
          example: "2023-06-13"
        customer_id:
          type: integer 
          example: 1
        dealer_id:
          type: integer 
          example: 1
        employee_id:
          type: integer
          example: 1 
        net_price:
          type: number 
          example: 100.0
        net_tax:
          type: number 
          example: 30.0
...
    StateTax:
      type: object
      properties:
        state:
          type: string 
          example: california
        tax: 
          type: number 
          example: 10 
    GetAllStateTaxResponse:
      type: object 
      properties:
        msg:
          type: array
          items:
            $ref: '#/components/schemas/StateTaxObject'
        status:
          type: string 
    StateTaxObject:
      type: object
      properties:
        state_tax_id:
          type: integer
          example: 1 
        state:
          type: string 
          example: california
        tax: 
          type: number 
          example: 10 
        created_at:
          type: string 
          example: Mon, 12 Jun 2022 08:36:01 GMT  
        modified_at:
          type: string 
          example: Mon, 12 Jun 2022 08:36:01 GMT
    GetStateTaxResponse:
      type: object 
      properties:
        msg:
          $ref: '#/components/schemas/StateTaxObject'
        status:
          type: string 
          example: success
    StateTaxPatch:
      type: object
      properties:
        tax: 
          type: number 
          example: 10 
    ApiResponse:
        type: object
        properties:
          msg:
            type: string
          status:
            type: string 